pyproject.toml
[project]
name = "todo-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.12",
    "psycopg[binary,pool]>=3.2.9",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.13",
    "tortoise-orm>=0.25.1",
]
